refactor: fold the rides screens into a single route
