--- FateConnect/Web/src/pages/Rides/screens/OfferRide/constants/index.ts
@@ -1,3 +1,0 @@
-export const OFFER_TITLE = 'Ofertar Nova Carona';
-export const OFFER_SUBTITLE = 'Compartilhe uma carona com outros estudantes da Fatec Sorocaba';
-export const OFFER_BUTTON_LABEL = 'Ofertar Carona';
